+++ src/client/components/Services/Services.css
@@ -1,4 +1,4 @@
-.service-details {
+.services {
     .body{
         margin: 16px;
     }
