+++ src/client/components/Services/TotalAmount/index.js
