+++ src/client/components/Services/index.js
@@ -1,0 +1,2 @@
+import Services from "./Services";
+export default Services;
