--- src/client/components/ViewServiceDetails/index.js
@@ -1,2 +1,0 @@
-import ViewServiceDetails from "./ViewServiceDetails";
-export default ViewServiceDetails;
